FIX: Fixed a bug where we threw away the first command in a file
It looks like this was done as a hack to drop the first line from the parsed output because parsing certain files were UTF-16 encoded and contained a BOM as the first character. Thus the first line always got parsed as a command and since they were often comments, it was ok to drop them.